backend/requirements.txt
2to3==1.0
absl-py==0.5.0
asn1crypto==0.24.0
astroid==2.0.4
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.3.1
grpcio==1.15.0
grpcio-tools==1.15.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mongoengine==0.15.3
pkg-resources==0.0.0
protobuf==3.6.1
pycparser==2.19
PyJWT==1.6.4
pylint==2.1.1
pymongo==3.7.1
six==1.11.0
typed-ast==1.1.0
virtualenv==16.0.0
wrapt==1.10.11
